feat(client): introduce compatibility for Service impl
